% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_exp_mixtures}
\alias{gene_exp_mixtures}
\title{Expression profiles of cell type mixtures}
\format{A matrix containing the gene expression profiles of simulated
  mixtures of cell types. Rows represent different genes. Columns represent
  differen mixture samples.}
\source{
The original gene expression profiles used to create these mixtures
  came from the deconvoluted expression profiles generated by applying EDec
  to the TCGA breast tumor dataset. These profiles can be found at:
  \link{http://genboree.org/theCommons/documents/571}. The original profiles
  can also be found in: \code{\link{true_cell_type_gene_exp}}.
}
\usage{
gene_exp_mixtures
}
\description{
A matrix containing simulated mixtures of gene expression profiles of 6
different cell types. There are two types of mixtures: those representing
methylation profiles of breast tumor samples (200), and those representing
normal breast samples (100).
}
\details{
Proportions of constituent cell types used to generate these 300 expression
profiles were the same ones used to generate the 300 methylation profiles in
\code{\link{meth_mixtures}}, so that for each expression profile there is a
matching methylation profile. Proportions of constituent cell types in the
200 samples representing breast tumors were generated from a Dirichlet
distribution with parameters that led to mean proportions of 59% cancer
epithelial, 9% cancer stroma, 19% cancer immune, 1% normal epithelial, 1%
normal stroma, and 1% normal immune. Proportions of constituent cell types in
the 100 samples representing normal breast were generated from a Dirichlet
distribution with parameters that led to mean proportions of 1% cancer
epithelial, 1% cancer stroma, 1% cancer immune, 49% normal epithelial, 39%
normal stroma, and 9% normal immune. In each mixture sample, independent
log-normal random variables were used to generate noisy versions of the
expression profiles of each constituent cell type. Once this procedure was
performed for every locus of all expression profiles that would constitute a
particular mixture sample, a linear combination of the noisy expression
profiles with the set of proportions of each constituent cell type generated
for that sample gave us the expression profile of that mixture.
}
\keyword{datasets}

