% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meth_mixtures}
\alias{meth_mixtures}
\title{Methylation profiles of cell type mixtures}
\format{A matrix containing the methylation profiles of simulated mixtures of
  cell types. Rows represent Illumina Infinium HumanMethylation27 BeadChip
  probes. Columns represent different mixture samples.}
\source{
The original methylation profiles used to create the mixtures in this
  dataset were gathered from different GEO datasets: GSE44837, GSE40699,
  GSE35069, GSE39981, GSE40699, GSE16368. We used different methylation
  profiles to represent different cell types in this simulated dataset.
  Cancer epithelial cell type methylation profile came from HCC38 cell line.
  Normal epithelial profile came from purified breast luminal epithelial
  cells. The cancer stromal profile came from IMR90 cell line. The normal
  stromal profile came from AG04449 cell line. The cancer immune came from
  purified CD8+ T-cells. The normal immune came from purified monocytes. The
  original profiles can also be found in: \code{\link{true_cell_type_meth}}
}
\usage{
meth_mixtures
}
\description{
A matrix containing simulated mixtures of experimentally derived methylation
profiles of 6 different cell types. There are two types of mixtures: those
representing methylation profiles of breast tumor samples (200), and those
representing normal breast samples (100).
}
\details{
The mixture samples in this dataset were created by mixing the methylation
profiles of representing cancer epithelial, cancer stroma, cancer immune,
normal epithelial, normal stroma, and normal immune cell types. Proportions
of constituent cell types used to generate these 300 methylation profiles
were the same ones used to generate the 300 expression profiles in
\code{\link{gene_exp_mixtures}}, so that for each methylation profile there
is a matching expression profile. Proportions of constituent cell types in
the 200 samples representing breast tumors were generated from a Dirichlet
distribution with parameters that led to mean proportions of 59% cancer
epithelial, 9% cancer stroma, 19% cancer immune, 1% normal epithelial, 1%
normal stroma, and 1% normal immune. Proportions of constituent cell types in
the 100 samples representing normal breast were generated from a Dirichlet
distribution with parameters that led to mean proportions of 1% cancer
epithelial, 1% cancer stroma, 1% cancer immune, 49% normal epithelial, 39%
normal stroma, and 9% normal immune. In each mixture sample, independent beta
random variables were used to generate noisy versions of the methylation
profiles of each constituent cell type. Once this procedure was performed for
every locus of all methylation profiles that would constitute a particular
mixture sample, a linear combination of the noisy methylation profiles with
the set of proportions of each constituent cell types generated for that
sample gave us the methylation profile of that mixture.
}
\keyword{datasets}

