% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_meth}
\alias{reference_meth}
\title{Reference methylation profiles}
\format{A matrix containing methylation profiles for various cell types. Rows
  represent Illumina Infinium HumanMethylation27 BeadChip probes. Columns
  represent different cell types.}
\source{
These methylation profiles were gathered from different GEO datasets:
  GSE44837, GSE40699, GSE35069, GSE39981, GSE40699, GSE16368.
}
\usage{
reference_meth
}
\description{
This dataset contains methylation profiles for four different classes of cell
types: breast cancer cell lines, purified normal breast epithelial cells,
stromal cells, and immune cells.
}
\keyword{datasets}

