% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_meth_class}
\alias{reference_meth_class}
\title{Cell type classes of reference methylation profiles}
\format{A vector of 4 possible strings: "cancer_ep", "normal_ep", "stromal",
  and "immune". Each string represents a different class of cell type. Each
  element in this vector corresponds to a column in
  \code{\link{reference_meth}}.}
\usage{
reference_meth_class
}
\description{
This dataset contains the cell type class associated with each of the
reference methylation profiles in the \code{\link{reference_meth}} dataset.
}
\keyword{datasets}

