% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{true_cell_type_gene_exp}
\alias{true_cell_type_gene_exp}
\title{True cell type specific gene expression}
\format{A matrix containing the gene expression profiles of each cell type.
  Rows represent different genes. Columns represent different cell types.}
\usage{
true_cell_type_gene_exp
}
\description{
This dataset contains the gene expression profiles of each constituent cell
type used to build the \code{\link{gene_exp_mixtures}} dataset.
}
\keyword{datasets}

