% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{true_cell_type_meth}
\alias{true_cell_type_meth}
\title{True cell type specific methylation}
\format{A matrix containing the methylation profiles of each cell type. Rows
  represent Illumina Infinium HumanMethylation27 BeadChip probes. Columns
  represent different cell types.}
\usage{
true_cell_type_meth
}
\description{
This dataset contains the methylation profiles of each constituent cell
type used to build the \code{\link{meth_mixtures}} dataset.
}
\keyword{datasets}

