% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{true_cell_type_props}
\alias{true_cell_type_props}
\title{True cell type proportions}
\format{A matrix containing proportions of constituent cell types in each
  mixture sample. Rows represent different samples. Columns represent
  different cell types.}
\usage{
true_cell_type_props
}
\description{
This dataset contains the cell type proportions used to build the 300
mixtures in both the \code{\link{meth_mixtures}} and the
\code{\link{gene_exp_mixtures}} datasets.
}
\keyword{datasets}

